requirements.txt
langchain~=0.3.14
langchain-ollama
langchain_openai

python-dotenv~=1.0.1
httpx~=0.27.2

streamlit~=1.40.1
requests~=2.32.3
pydantic~=2.10.5